+++ lib/model/report_reg_verticality.dart
@@ -19,7 +19,7 @@ class ReportRegVerticality {
   late String theodolite1;
   late String theodolite2;
   late String alatUkur;
-  late String toleransiKetegakan;
+  late int toleransiKetegakan;
   final List<ValueVerticality>? valueVerticality;
 
   ReportRegVerticality({